requirements.txt
jpype1>=1.5.2